move update_checker to top-level
